separate avi creation script to new module
